LayoutTests/imported/w3c/web-platform-tests/navigation-api/updateCurrentEntry-method/same-document-away-and-back-location-api-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected 3 but got 2

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 3 but got 2

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected 3 but got 2

TIMEOUT entry.getState() behavior after navigating away using the location API, then back Test timed out
